chore: jwt auth and restructure
- implement jwt based auth
- restructure project to avoid circular dependency
- cleanup

+++ internal/user/user_types.go
@@ -3,7 +3,7 @@ package user
 import "time"
 
 type User struct {
-	Id             int        `db:"id" json:"id"`
+	Id             int64      `db:"id" json:"id"`
 	FirstName      string     `db:"firstName" json:"firstName"`
 	LastName       string     `db:"lastName" json:"lastName"`
 	Username       string     `db:"username" json:"username"`
